requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
attrs==26.1.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
chromadb==1.1.1
click==8.1.8
cloudpickle==3.1.2
contourpy==1.3.3
crewai==1.11.0
cryptography==46.0.5
curl_cffi==0.13.0
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
et_xmlfile==2.0.0
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
inflection==0.5.1
instructor==1.14.5
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
json5==0.10.0
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kubernetes==35.0.0
lance-namespace==0.6.1
lance-namespace-urllib3-client==0.6.1
lancedb==0.30.1
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.22
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
linkify-it-py==2.1.0
llvmlite==0.44.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mcp==1.26.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mmh3==5.2.1
more-itertools==10.8.0
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.7.1
multitasking==0.0.12
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.24.4
openai==2.29.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==3.0.1
pandas-datareader==0.10.0
pandas-ta==0.4.71b0
pdfminer.six==20251230
pdfplumber==0.11.9
peewee==4.0.2
pillow==12.1.1
pip-tools==7.5.3
platformdirs==4.9.4
portalocker==2.7.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.12.1
pyparsing==3.3.2
pypdf==6.9.2
pypdfium2==5.6.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.22
pytz==2026.1.post1
PyWavelets==1.9.0
PyYAML==6.0.3
Quandl==3.7.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==82.0.1
shap==0.51.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sse-starlette==3.3.3
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
textual==8.1.1
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.0.2
tomli_w==1.1.0
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.23.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==2.0.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uv==0.9.30
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wheel==0.46.3
xgboost==3.2.0
xxhash==3.6.0
yarl==1.23.0
yfinance==1.2.0
youtube-transcript-api==1.2.4
yt-dlp==2026.3.17
zipp==3.23.0
zstandard==0.25.0
